WIP: change remote to write image OS based on base images, or options
- Change remote to init empty images with OS/Arch to linux/amd64,
but override based on any Base or Previous image is supplied, or
when explicity set by options. This preserves current behavior
but gives overrides for using windows images.

- Image format does not default to GOOS, though it was considered.
But the inconsistent behavior might be problematic, particularly
when `pack` and `lifecycle` might be different binaries but used
in the same user interaction (e.g. `pack` on MacOS, building
Windows images with Windows-compiled lifecycle binaries)

- Change remote to disambiguate when image lists contain windows/linux images.
It picks based on current image format, base image or explicit option, if present.

- Refactor test methods to use ggcr directly; simpler queries for OS