Add default PAX record permissons to Windows layer entries
- Set default PAX-record file permissions (`rawsd` field) when none are present
- Set Owner to BUILTIN\Administrators and Group to Administrators when Uid *and* Gid are 0
- Set Owner to BUILTIN\Users and Group to Users otherwise

This is needed for images to run on k8s (ContainerD) containers. If a layer entries is created w/o PAX records, the are inaccessible to `USER`. Docker appears to workaround entries w/o PAX records by making them owned by the run `USER`.

This implementation infers PAX record permissions based on the GID/UID, only when the PAX record permissions is not already set. The syntax for the `rawsd` field is the base64 encoded binary format of the security descriptor from the SDDLs in the comments.

We attempt to map the intent of the UID/GID to windows:
- Both UID and GID are 0 => Owner + Primary Group is BUILTIN\Administrators
- Any other values => Owner + Primary Group is BUILTIN\Users (default group for any real user on the system)

This is a basic implementaion to start with, but a more fine-grained mapping will likely be preferable, perhaps informed by how Linux NTFS or SMB drivers map permissions.

Signed-off-by: Micah Young <ymicah@vmware.com>