Changes registry test helper to support non-localhost
This is useful for consumers of DockerRegistry helper, particularly lifecycle and pack, when running test containers that access the separate registry container, listening on a published host port (currently only possible with `--network=host` which can't be used on Windows). This also lets a Unix workstation run tests against a remote Windows daemon with `DOCKER_HOST=tcp://<daemon host>` set.

The logic to detect the host IP:
  - Any DOCKER_HOST=tcp://<host>
  - Any `host.docker.internal` that resolves, use the IP
  - Any exact IPs (`<host IP>/32`) in `insecure-registry` entries
  - ... or, fallback to "localhost"

This does not change the current behavior for imgutil when running on GHA, which has none of the above set so will use "localhost".

This also replaces `stefanscherer/registry-windows`, which had a bug where it would return sporadic 503s. This bug appears fixed in docker-distribution 2.7.1 but the image hasn't been update for over 2 years. This is replaced with a newly created [`micahyoung/registry`](https://github.com/micahyoung/registry-image) which hosts equivalent Windows & Linux AMD64/ARM images in a single manifest list.

Signed-off-by: Micah Young <ymicah@vmware.com>