Use DriverStatus to check if containerd is used
This change looks at the DriverStatus that contains a "driver-type"
equal to "io.containerd.snapshotter.v1" if the daemon uses containerd
for storing images, this is slightly better than looking at the Driver.
The Driver is "stargz" by default on Docker Desktop but a user can change
this default value, they could use "overlayfs" for example, or any other
containerd snapshotter.

Signed-off-by: Djordje Lukic <djordje.lukic@docker.com>