sparse: do not modify media type when unspecified
The media type should only be modified when the user specifies it. In my case I noticed the digest given by the sparse.Image was not the same as the image it was dervied from. This was because the logic to set the media type defaulted to overriding to OCI if the media type was not specified. This commit changes the logic to only override the media type if the user specifies it.

Signed-off-by: Jesse Brown <jabrown85@gmail.com>