Ensure that downloadable layers are always downloaded
Previously, we forced the layer to be downloaded by asking for its size.
But, this has performance implications when you only need the size to construct a manifest.
Now, we force the layer to be downloaded by opening a ReadCloser to the layer,
and size only reports if the layer is present on disk, it doesn't cause the layer to be downloaded.

Signed-off-by: Natalie Arellano <narellano@vmware.com>