imgutil(local): Add Logger functionality(#282)
This commit enhances the Logger functionality by integrating it with the existing
automatic detection of the storage driver type in the NewStore function.
The Logger interface has been added to log warnings related to the containerd storage driver.
Test cases have been created to ensure that warnings are logged correctly when using containerd.

Signed-off-by: wuhenggongzi <1305678586@qq.com>